requirements.txt
Babel==2.1.1
Django==1.9
amqp==1.4.8
anyjson==0.3.3
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
billiard==3.3.0.22
celery==3.1.19
certifi==2015.11.20.1
django-cache-machine==0.9.1
django-celery==3.1.17
django-cors-headers==1.1.0
django-debug-toolbar==1.4
djangorestframework==3.3.2
flower==0.8.3
futures==3.0.3
kombu==3.0.30
librabbitmq==1.6.1
lxml==3.5.0
psycopg2==2.6.1
pylibmc==1.5.0
pytz==2015.7
redis==2.10.5
requests==2.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.1.18
tornado==4.3
wsgiref==0.1.2
